Polish SEO metadata and internal structure
